Fix Windows packaging prep and calm start script heuristics
