+++ .gitignore
@@ -25,5 +25,6 @@ Thumbs.db
 /.venv/
 %LOCALAPPDATA%/
 apps/arw-launcher/src-tauri/gen/
+apps/arw-launcher/src-tauri/bin/
 __pycache__/
 *.pyc
